Move `CommandLineTool` from `core-utils` to `common-utils`
This removes the warning log when a version requirement is not satisfied
and the user configured to ignore this, but that should be fine in the
context of upcoming changes that will relax the version requirements on
external tools.
